Implementado a impressão do grafo e reformulado a função insere.

+++ 2 - Grafo - Lista de Adjacência/Grafos não direcionados/0 - Operações/grafo_listaadj_struct.h
@@ -3,7 +3,9 @@
 #define VERTICE_INVALIDO NULL
 #define AN -1
 
-typedef int Peso;
+//typedef int Peso;
+
+typedef float Peso;
 
 /*
 	Uma estrutura do tipo "Aresta" é mais interessante para este
